Collect some common reduce and map operations to reuse them easier
